requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-genai
deepgram-sdk==3.3.2
motor
python-multipart
websockets